+++ HISTORY.md
@@ -6,10 +6,13 @@ All notable changes to this project will be documented in this file.
 
 <!-- add unreleased items here -->
 
+* Fixed
+  * Log|Info|Debug output are properly formatted (via [#165])
 * Tests
   * Added testbeds for integration (via [#164])
 
 [#164]: https://github.com/CycloneDX/cyclonedx-esbuild/pull/164
+[#165]: https://github.com/CycloneDX/cyclonedx-esbuild/pull/165
 
 ## 1.4.0 - 2026-06-29